added one to many relationship between users and medication

+++ medication/src/main/java/com/zhuk/coursework/service/MedicationService.java
@@ -4,18 +4,25 @@ import com.zhuk.coursework.dto.MedicationDto;
 import com.zhuk.coursework.dto.NewMedicationDto;
 import com.zhuk.coursework.dto.UpdateQuantityDto;
 import com.zhuk.coursework.entity.MedicationEntity;
+import com.zhuk.coursework.entity.UserEntity;
 import com.zhuk.coursework.enums.ApiMessageEnum;
 import com.zhuk.coursework.enums.ErrorCodeEnum;
 import com.zhuk.coursework.enums.MedicationTypeEnum;
+import com.zhuk.coursework.enums.UserRoleEnum;
 import com.zhuk.coursework.exception.medication.MedicationAlreadyExistsException;
+import com.zhuk.coursework.exception.medication.MedicationNoRightsException;
 import com.zhuk.coursework.exception.medication.MedicationNotFoundException;
 import com.zhuk.coursework.exception.medication.NotEnoughQuantityException;
+import com.zhuk.coursework.exception.user.UserUnknownException;
 import com.zhuk.coursework.mapper.MedicationMapper;
 import com.zhuk.coursework.repository.MedicationRepository;
+import com.zhuk.coursework.security.CustomUserDetails;
 import com.zhuk.coursework.utils.ErrorCodeHelper;
 import com.zhuk.coursework.utils.MessageSourceWrapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -29,14 +36,16 @@ public class MedicationService {
     private final MedicationRepository medicationRepository;
     private final MessageSourceWrapper messageSourceWrapper;
     private final ErrorCodeHelper errorCodeHelper;
-
+    private final UserService userService;
     public List<MedicationDto> findAll() {
-        return medicationRepository.findAll().stream()
+        List<MedicationEntity> medicationEntities = getMedicationEntityListForCurrentUser();
+        return medicationEntities.stream()
                 .map(medicationMapper::map)
                 .toList();
     }
 
     public MedicationDto findById(Long id) {
+        validateMedicationForCurrentUser(id);
         return medicationMapper.map(getEntityByIdOrThrowException(id));
     }
 
@@ -44,14 +53,27 @@ public class MedicationService {
         if(getOptionalEntityByNameAndWeight(dto.getName(), dto.getWeight()).isPresent()) {
             throw new MedicationAlreadyExistsException(HttpStatus.BAD_REQUEST,
                     messageSourceWrapper.getMessageCode(ApiMessageEnum.MEDICATION_ALREADY_EXISTS),
-                    errorCodeHelper.getCode(ErrorCodeEnum.MEDICATION_ALREADY_EXISTS));
+                    errorCodeHelper.getCode(ErrorCodeEnum.MEDICATION_ALREADY_EXISTS_CODE));
         }
-        MedicationEntity entity = medicationMapper.mapFromNewDto(dto);
+        MedicationEntity entity = MedicationEntity.builder()
+                .name(dto.getName())
+                .type(MedicationTypeEnum.valueOf(dto.getType()))
+                .additionalInfo(dto.getAdditionalInfo())
+                .manufacturer(dto.getManufacturer())
+                .quantity(dto.getQuantity())
+                .weight(dto.getWeight())
+                .user(userService.getUserEntityByUsernameOrThrowException(getCurrentUser().getUsername()))
+                .build();
         return medicationMapper.map(medicationRepository.save(entity));
     }
 
     @Transactional
     public void deleteMedication(Long id) {
+        Optional<MedicationEntity> entityOptional = getOptionalEntityById(id);
+        if(entityOptional.isEmpty()) {
+            return;
+        }
+        validateMedicationForCurrentUser(id);
         medicationRepository.deleteById(id);
     }
 
@@ -61,11 +83,31 @@ public class MedicationService {
         if(entityOptional.isEmpty()) {
             return saveMedication(dto);
         }
-        updateByNameAndWeight(dto);
         Long id = entityOptional.get().getId();
-        MedicationEntity entity = medicationMapper.mapFromNewDto(dto).toBuilder().id(id).build();
+        validateMedicationForCurrentUser(id);
+        updateByNameAndWeight(dto);
+        MedicationEntity entity = MedicationEntity.builder()
+                .id(id)
+                .type(MedicationTypeEnum.valueOf(dto.getType()))
+                .additionalInfo(dto.getAdditionalInfo())
+                .manufacturer(dto.getManufacturer())
+                .quantity(dto.getQuantity())
+                .weight(dto.getWeight())
+                .user(userService.getUserEntityByUsernameOrThrowException(getCurrentUser().getUsername()))
+                .build();
         return medicationMapper.map(entity);
+    }
 
+    @Transactional
+    public void updateQuantity(Long id, UpdateQuantityDto dto) {
+        MedicationEntity entity = getEntityByIdOrThrowException(id);
+        validateMedicationForCurrentUser(id);
+        if(dto.getQuantity() < 0 && entity.getQuantity() < Math.abs(dto.getQuantity())) {
+            throw new NotEnoughQuantityException(HttpStatus.BAD_REQUEST,
+                    messageSourceWrapper.getMessageCode(ApiMessageEnum.MEDICATION_NOT_ENOUGH_QUANTITY),
+                    errorCodeHelper.getCode(ErrorCodeEnum.MEDICATION_NOT_ENOUGH_QUANTITY_CODE));
+        }
+        medicationRepository.updateQuantityById(id, entity.getQuantity() + dto.getQuantity());
     }
 
     private void updateByNameAndWeight(NewMedicationDto dto) {
@@ -79,7 +121,7 @@ public class MedicationService {
         return medication.orElseThrow(
                 () -> new MedicationNotFoundException(HttpStatus.NOT_FOUND,
                         messageSourceWrapper.getMessageCode(ApiMessageEnum.MEDICATION_NOT_FOUND),
-                        errorCodeHelper.getCode(ErrorCodeEnum.MEDICATION_NOT_FOUND))
+                        errorCodeHelper.getCode(ErrorCodeEnum.MEDICATION_NOT_FOUND_CODE))
         );
     }
 
@@ -91,14 +133,31 @@ public class MedicationService {
         return medicationRepository.getMedicationEntityByNameAndWeight(name, weight);
     }
 
-    @Transactional
-    public void updateQuantity(Long id, UpdateQuantityDto dto) {
-        MedicationEntity entity = getEntityByIdOrThrowException(id);
-        if(dto.getQuantity() < 0 && entity.getQuantity() < Math.abs(dto.getQuantity())) {
-            throw new NotEnoughQuantityException(HttpStatus.BAD_REQUEST,
-                    messageSourceWrapper.getMessageCode(ApiMessageEnum.NOT_ENOUGH_QUANTITY),
-                    errorCodeHelper.getCode(ErrorCodeEnum.NOT_ENOUGH_QUANTITY));
+    private List<MedicationEntity> getMedicationEntityListForCurrentUser() {
+        UserEntity user = userService.getUserEntityByUsernameOrThrowException(getCurrentUser().getUsername());
+        if(user.getRole() == UserRoleEnum.ROLE_ADMIN) {
+            return medicationRepository.findAll();
         }
-        medicationRepository.updateQuantityById(id, entity.getQuantity() + dto.getQuantity());
+        return user.getMedications();
+    }
+
+    private CustomUserDetails getCurrentUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)) {
+            throw new UserUnknownException(HttpStatus.FORBIDDEN,
+                    messageSourceWrapper.getMessageCode(ApiMessageEnum.USER_UNKNOWN_EXCEPTION),
+                    errorCodeHelper.getCode(ErrorCodeEnum.USER_UNKNOWN_EXCEPTION_CODE));
+        }
+        return ((CustomUserDetails) authentication.getPrincipal());
+    }
+
+    private void validateMedicationForCurrentUser(Long medicationId) {
+        MedicationEntity medicationEntity = getEntityByIdOrThrowException(medicationId);
+        UserEntity user = userService.getUserEntityByUsernameOrThrowException(getCurrentUser().getUsername());
+        if(!medicationEntity.getUser().equals(user)) {
+            throw new MedicationNoRightsException(HttpStatus.FORBIDDEN,
+                    messageSourceWrapper.getMessageCode(ApiMessageEnum.MEDICATION_NO_RIGHTS),
+                    errorCodeHelper.getCode(ErrorCodeEnum.MEDICATION_NO_RIGHTS_CODE));
+        }
     }
 }
